Update UKF, Tests, and Other
Updated the UKF since there were some bugs in the interface. Also
slightly modified the Cholesky, DiagonalSolver, and LDL functionality
to not have an implicit size_t to int warning.

+++ src/LinearAlgebra/Factorizations/Cholesky.hpp
@@ -47,7 +47,7 @@ void cholesky( const Matrix<T,C> & m, LMat<T> & L ){
     if (m.size() != L.size() ) {
         L.resize(m.size());
     }
-    const int numr = L.size().rows;
+    const int numr = static_cast<int>(L.size().rows);
     
     double sum = T();
     for (int i = 0; i < numr; i++) {
